Added link to delete all cached files of an application.
Refs: #95

+++ myfuses_tools/application/actDeletePath.php
@@ -1,11 +1,21 @@
 <?php
+/**
+ * MyFuses Framework (http://myfuses.candango.org)
+ *
+ * @link      http://github.com/candango/myfuses
+ * @copyright Copyright (c) 2006 - 2018 Flavio Garcia
+ * @license   https://www.apache.org/licenses/LICENSE-2.0  Apache-2.0
+ */
+
+use Candango\MyFuses\Util\FileHandler;
+
 function delpath($path)
 {
     if (is_dir($path)) {
         $it = new RecursiveDirectoryIterator($path);
         foreach (new RecursiveIteratorIterator($it, 1) as $child) {
             $pName = "" . $child;
-            if ($child->isDir() && !$child->isLink()) {
+            if ($child->isDir() && !$child->isLink() && file_exists($child)) {
                 rmdir($pName);
             } else {
                 unlink($pName);
@@ -17,8 +27,8 @@ function delpath($path)
     }  
 }
 
-$cachedPath = $application->getController()->getParsedPath() .
-    $application->getName();
+$cachedPath = FileHandler::sanitizePath(
+    $application->getController()->getParsedPath()) . $application->getName();
 
 // checking if the file delete was in the cached path
 if (strpos($_GET['file'], $cachedPath) !== false) {
